updated type annotations to python 3.10 style, modularized code
