+++ .pre-commit-config.yaml
@@ -1,27 +1,29 @@
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v2.3.0
+    rev: v4.4.0
     hooks:
       - id: check-yaml
       - id: end-of-file-fixer
       - id: trailing-whitespace
-  - repo: https://github.com/psf/black
-    rev: 19.3b0
+  -   repo: local
+      hooks:
+      # NOTE: we make `black` a local hook because if it's installed from
+      # PyPI (rather than from source) then it'll run twice as fast thanks to mypyc
+      -   id: black
+          name: black
+          description: "Black: The uncompromising Python code formatter"
+          entry: black
+          language: python
+          require_serial: true
+          types_or: [python, pyi]
+          additional_dependencies: [black==23.3.0]
+  - repo: https://github.com/charliermarsh/ruff-pre-commit
+    # Ruff version.
+    rev: 'v0.0.257'
     hooks:
-      - id: black
-  - repo: https://github.com/PyCQA/isort
-    rev: 5.9.3
-    hooks:
-      - id: isort
+      - id: ruff
   - repo: https://github.com/pre-commit/mirrors-mypy
     rev: v0.910 # Use the sha / tag you want to point at
     hooks:
       - id: mypy
         args: [--no-strict-optional, --ignore-missing-imports]
-  - repo: https://github.com/nbQA-dev/nbQA
-    rev: 1.1.1
-    hooks:
-      - id: nbqa-black
-      - id: nbqa-pyupgrade
-        args: [--py36-plus]
-      - id: nbqa-isort
