+++ pyproject.toml
@@ -1,21 +1,18 @@
-[tool.poetry]
-name = "ZenTables"
-version = "0.2.0b3"
+[project]
+name = "zentables"
+version = "0.3.0b1"
 description = "Stress-free descriptive tables in Python."
 authors = [
-    "Paul Xu <yang_xu@brown.edu>",
-    "Edward Huh <edward_huh@brown.edu>",
-    "Kevin Wilson <kevin_wilson@brown.edu>",
-    "The Policy Lab <thepolicylab@brown.edu>"
+    { name = "Paul Xu", email = "yang_xu@brown.edu" },
+    { name = "Edward Huh", email = "edward_huh@brown.edu" },
+    { name = "Kevin Wilson", email = "kevin_wilson@brown.edu" },
+    { name = "The Policy Lab", email = "thepolicylab@brown.edu" },
 ]
-license = "MIT"
-packages = [
-    { include = "zentables", from = "src" }
-]
+maintainers = [{ name = "Paul Xu", email = "yang_xu@brown.edu" }]
+requires-python = ">=3.9"
 readme = "README.md"
-repository = "https://github.com/thepolicylab/ZenTables.git"
 keywords = ["pandas", "tables", "display", "styler"]
-classifiers=[
+classifiers = [
     'Development Status :: 4 - Beta',
 
     'Intended Audience :: Science/Research',
@@ -23,48 +20,178 @@ classifiers=[
 
     'License :: OSI Approved :: MIT License',
 
-    'Programming Language :: Python :: 3',
-    'Programming Language :: Python :: 3.7',
-    'Programming Language :: Python :: 3.8',
-    'Programming Language :: Python :: 3.9'
+    'Programming Language :: Python :: 3.9',
+    'Programming Language :: Python :: 3.10',
+    'Programming Language :: Python :: 3.11',
+]
+dependencies = ["pandas>=2.0.0rc1"]
+
+[project.urls]
+repository = "https://github.com/thepolicylab/ZenTables.git"
+
+[project.optional-dependencies]
+dev = ["black[jupyter]", "ruff", "pytest", "pytest-asyncio", "mypy", "jupyterlab", "pre-commit"]
+
+[tool.hatch.envs.dev]
+python = "3.9"
+features = ["dev"]
+
+[tool.black]
+target-version = ['py39']
+required-version = '23.3.0'
+exclude = '''
+(
+    asv_bench/env
+  | \.egg
+  | \.git
+  | \.hg
+  | \.mypy_cache
+  | \.nox
+  | \.tox
+  | \.venv
+  | _build
+  | buck-out
+  | build
+  | dist
+  | setup.py
+)
+'''
+
+[tool.ruff]
+line-length = 88
+update-check = false
+target-version = "py39"
+fix = true
+unfixable = ["E711"]
+
+select = [
+    # pyflakes
+    "F",
+    # pycodestyle
+    "E",
+    "W",
+    # flake8-2020
+    "YTT",
+    # flake8-bugbear
+    "B",
+    # flake8-quotes
+    "Q",
+    # pylint
+    "PLE",
+    "PLR",
+    "PLW",
+    # misc lints
+    "PIE",
+    # flake8-pyi
+    "PYI",
+    # tidy imports
+    "TID",
+    # implicit string concatenation
+    "ISC",
+    # type-checking imports
+    "TCH",
 ]
 
-[tool.poetry.dependencies]
-python = "^3.7.1"
-pandas = "^1.3.0"
-Jinja2 = "^3.0.1"
 
-[tool.poetry.dev-dependencies]
-black = "^20.8b1"
-pylint = "^2.6.0"
-pytest = "^6.2.1"
-isort = "^5.7.0"
-ipdb = "^0.13.4"
-pytest-asyncio = "^0.14.0"
-nb-black = "^1.0.7"
-jupyterlab = "^3.0.16"
-mypy = "^0.910"
+ignore = [
+    # space before : (needed for how black formats slicing)
+    # "E203",  # not yet implemented
+    # module level import not at top of file
+    "E402",
+    # do not assign a lambda expression, use a def
+    "E731",
+    # line break before binary operator
+    # "W503",  # not yet implemented
+    # line break after binary operator
+    # "W504",  # not yet implemented
+    # controversial
+    "B006",
+    # controversial
+    "B007",
+    # controversial
+    "B008",
+    # setattr is used to side-step mypy
+    "B009",
+    # getattr is used to side-step mypy
+    "B010",
+    # tests use assert False
+    "B011",
+    # tests use comparisons but not their returned value
+    "B015",
+    # false positives
+    "B019",
+    # Loop control variable overrides iterable it iterates
+    "B020",
+    # Function definition does not bind loop variable
+    "B023",
+    # Functions defined inside a loop must not use variables redefined in the loop
+    # "B301",  # not yet implemented
+    # Too many arguments to function call
+    "PLR0913",
+    # Too many returns
+    "PLR0911",
+    # Too many branches
+    "PLR0912",
+    # Too many statements
+    "PLR0915",
+    # Global statements are discouraged
+    "PLW0603",
+    # Docstrings should not be included in stubs
+    "PYI021",
+    # Use typing_extensions.TypeAlias for type aliases
+    # "PYI026",  # not yet implemented
+    # Use "collections.abc.*" instead of "typing.*" (PEP 585 syntax)
+    # "PYI027",  # not yet implemented
+    # while int | float can be shortened to float, the former is more explicit
+    # "PYI041",  # not yet implemented
 
-[tool.isort]
-multi_line_output = 3
-include_trailing_comma = true
-force_grid_wrap = 0
-use_parentheses = true
-ensure_newline_before_comments = true
-line_length = 88
+    # Additional checks that don't pass yet
+    # Within an except clause, raise exceptions with ...
+    "B904",
+    # Magic number
+    "PLR2004",
+    # Consider `elif` instead of `else` then `if` to remove indentation level
+    "PLR5501",
+]
 
-[tool.pylint.basic]
-good-names = "i,j,k,ex,Run,_,df,pc"
-
-[tool.pylint.messages_control]
-disable = "C0330, C0326, R0912, R0913, R0914, R0915"
-
-[tool.pylint.format]
-max-line-length = "88"
+exclude = [".github", "docs", "src/notebook"]
 
 [tool.mypy]
 ignore_missing_imports = true
+python_version = "3.9"
+
+[tool.isort]
+known_pre_libs = "pandas._config"
+known_pre_core = [
+    "pandas._libs",
+    "pandas._typing",
+    "pandas.util._*",
+    "pandas.compat",
+    "pandas.errors",
+]
+known_dtypes = "pandas.core.dtypes"
+known_post_core = ["pandas.tseries", "pandas.io", "pandas.plotting"]
+sections = [
+    "FUTURE",
+    "STDLIB",
+    "THIRDPARTY",
+    "PRE_LIBS",
+    "PRE_CORE",
+    "DTYPES",
+    "FIRSTPARTY",
+    "POST_CORE",
+    "LOCALFOLDER",
+]
+profile = "black"
+combine_as_imports = true
+force_grid_wrap = 2
+force_sort_within_sections = true
+skip_glob = "env"
+skip = "pandas/__init__.py"
 
 [build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build]
+sources = ["src/zentables"]
